added move ticket operation, still need to write the new submit to save the new board id to move the tickets
