+++ client/src/components/bulk-actions/BulkActionsFormStep2.tsx
@@ -1,16 +1,16 @@
 import React from "react"
-import { BulkEditOption, BulkEditOptionKey } from "./BulkActionsForm"
+import { BulkEditOperation, BulkEditOperationKey } from "./BulkActionsForm"
 
 interface Props {
 	step: number
 	setStep: (step: number) => void
 	numSelectedIssues: number
-	operation: BulkEditOptionKey | null | undefined
-	setOperation: (option: BulkEditOptionKey | null | undefined) => void
+	operation: BulkEditOperationKey | null | undefined
+	setOperation: (operation: BulkEditOperationKey | null | undefined) => void
 }
 
-export const BulkActionsFormStep2 = ({step, setStep, numSelectedIssues, operation, setOperation}: Props) => {
-	const options: Array<BulkEditOption> = [
+export const BulkActionsFormStep2 = ({step, setStep, numSelectedIssues, operation: propOperation, setOperation}: Props) => {
+	const operations: Array<BulkEditOperation> = [
 		{
 			key: "edit-issues",
 			text: "Edit Issues",
@@ -42,24 +42,24 @@ export const BulkActionsFormStep2 = ({step, setStep, numSelectedIssues, operatio
 			<h3 className = "tw-m-0">Step 2 of 4: Choose Operation</h3>
 			<p>Choose the operation you wish to perform on <span className = "tw-font-semibold">{numSelectedIssues}</span> issues</p>
 			<div>
-				{options.map((option, index) => {
+				{operations.map((operation, index) => {
 					return (
-						<div key={option.key} className = {`${index === 0 ? "tw-border-y" : "tw-border-b"} tw-flex tw-flex-row tw-py-2 tw-gap-x-4 tw-border-gray-300`} >
+						<div key={operation.key} className = {`${index === 0 ? "tw-border-y" : "tw-border-b"} tw-flex tw-flex-row tw-py-2 tw-gap-x-4 tw-border-gray-300`} >
 							<div>
-								<input onChange={() => setOperation(option.key as BulkEditOptionKey)} value={option.key} checked={operation === option.key} type = "radio"/>	
+								<input onChange={() => setOperation(operation.key as BulkEditOperationKey)} value={operation.key} checked={propOperation === operation.key} type = "radio"/>	
 							</div>
 							<div className = "tw-flex tw-flex-1">
-								<p>{option.text}</p>
+								<p>{operation.text}</p>
 							</div>
 							<div className = "tw-flex tw-flex-1">
-								<p>{option.description}</p>
+								<p>{operation.description}</p>
 							</div>
 						</div>
 					)
 				})}
 			</div>
 			<div className = "tw-flex tw-flex-row tw-gap-x-2">
-				<button onClick={() => setStep(step+1)} disabled={!operation} className = "button">Next</button>			
+				<button onClick={() => setStep(step+1)} disabled={!propOperation} className = "button">Next</button>			
 				<button onClick={() => setStep(step-1)} className = "button --secondary">Cancel</button>			
 			</div>
 		</div>
